Parallel-enabled hyperspace walker
