+++ h2o-core/src/main/java/hex/grid/hyperspace/HyperSpaceWalker.java
@@ -1,4 +1,4 @@
-package hex.grid;
+package hex.grid.hyperspace;
 
 import hex.Model;
 import hex.ModelParametersBuilderFactory;
@@ -14,6 +14,16 @@ import static java.lang.StrictMath.min;
 public interface HyperSpaceWalker<MP extends Model.Parameters, C extends HyperSpaceSearchCriteria> {
 
   interface HyperSpaceIterator<MP extends Model.Parameters> {
+
+    /**
+     * Prepares up to n initial model parameters to start a walk through hyperspace.
+     * May return a {@link List} with less model parameters than requested, or an empty list.
+     *
+     * @param numParams Number of models to retrieve
+     * @return A possibly empty {@link List} of Model.Parameters. Never null.
+     */
+    List<MP> initialModelParameters(int numParams);
+
     /**
      * Get next model parameters.
      *
@@ -24,21 +34,23 @@ public interface HyperSpaceWalker<MP extends Model.Parameters, C extends HyperSp
      * <p>The method can optimize based on previousModel, but should be
      * able to handle null-value.</p>
      *
-     * @param previousModel  model generated for the previous point in hyper space, can be null.
+     * @param previousModel  Optional model generated for any previous point in the hyperspace, not necessarily the last 
+     *                       parameters (time-wise) generated by the hyperspace walker.
+     *                       Presence of previously-built model is purely optional.
      *
-     * @return model parameters for next point in hyper space or null if there is no such point.
+     * @return An instance of Model.Parameter, if any.
      *
      * @throws IllegalArgumentException  when model parameters cannot be constructed
      * @throws java.util.NoSuchElementException if the iteration has no more elements
      */
-    MP nextModelParameters(Model previousModel);
+    MP nextModelParameters(Optional<Model> previousModel);
 
     /**
      * Returns true if the iterator can continue.  Takes into account strategy-specific stopping criteria, if any.
      * @param previousModel  optional parameter which helps to determine next step, can be null
      * @return  true if the iterator can produce one more model parameters configuration.
      */
-    boolean hasNext(Model previousModel);
+    boolean hasNext(Optional<Model> previousModel);
 
     void reset();
 
@@ -186,9 +198,9 @@ public interface HyperSpaceWalker<MP extends Model.Parameters, C extends HyperSp
         HyperSpaceSearchCriteria.Strategy strategy = search_criteria.strategy();
 
         if (strategy == HyperSpaceSearchCriteria.Strategy.Cartesian) {
-          return new HyperSpaceWalker.CartesianWalker<>(params, hyperParams, paramsBuilderFactory, (HyperSpaceSearchCriteria.CartesianSearchCriteria) search_criteria);
+          return new CartesianWalker<>(params, hyperParams, paramsBuilderFactory, (HyperSpaceSearchCriteria.CartesianSearchCriteria) search_criteria);
         } else if (strategy == HyperSpaceSearchCriteria.Strategy.RandomDiscrete ) {
-          return new HyperSpaceWalker.RandomDiscreteValueWalker<>(params, hyperParams, paramsBuilderFactory, (HyperSpaceSearchCriteria.RandomDiscreteValueSearchCriteria) search_criteria);
+          return new RandomDiscreteValueWalker<>(params, hyperParams, paramsBuilderFactory, (HyperSpaceSearchCriteria.RandomDiscreteValueSearchCriteria) search_criteria);
         } else {
           throw new H2OIllegalArgumentException("strategy", "GridSearch", strategy);
         }
@@ -272,7 +284,6 @@ public interface HyperSpaceWalker<MP extends Model.Parameters, C extends HyperSp
         }
       }
     } // BaseWalker()
-
     @Override
     public String[] getHyperParamNames() {
       return _hyperParamNames;
@@ -334,267 +345,4 @@ public interface HyperSpaceWalker<MP extends Model.Parameters, C extends HyperSp
       return Arrays.deepHashCode(hashMe);
     }
   }
-
-  /**
-   * Hyperparameter space walker which visits each combination of hyperparameters in order.
-   */
-  public static class CartesianWalker<MP extends Model.Parameters>
-          extends BaseWalker<MP, HyperSpaceSearchCriteria.CartesianSearchCriteria> {
-
-    public CartesianWalker(MP params,
-                           Map<String, Object[]> hyperParams,
-                           ModelParametersBuilderFactory<MP> paramsBuilderFactory,
-                           HyperSpaceSearchCriteria.CartesianSearchCriteria search_criteria) {
-      super(params, hyperParams, paramsBuilderFactory, search_criteria);
-    }
-
-    @Override
-    public HyperSpaceIterator<MP> iterator() {
-
-      return new HyperSpaceIterator<MP>() {
-        /** Hyper params permutation.
-         */
-        private int[] _currentHyperparamIndices = null;
-
-        @Override
-        public MP nextModelParameters(Model previousModel) {
-          _currentHyperparamIndices = _currentHyperparamIndices != null ? nextModelIndices(_currentHyperparamIndices) : new int[_hyperParamNames.length];
-          if (_currentHyperparamIndices != null) {
-            // Fill array of hyper-values
-            Object[] hypers = hypers(_currentHyperparamIndices, new Object[_hyperParamNames.length]);
-            // Get clone of parameters
-            MP commonModelParams = (MP) _params.clone();
-            // Fill model parameters
-            MP params = getModelParams(commonModelParams, hypers);
-
-            return params;
-          } else {
-            throw new NoSuchElementException("No more elements to explore in hyper-space!");
-          }
-        }
-
-        @Override
-        public boolean hasNext(Model previousModel) {
-          if (_currentHyperparamIndices == null) {
-            return true;
-          }
-          int[] hyperparamIndices = _currentHyperparamIndices;
-          for (int i = 0; i < hyperparamIndices.length; i++) {
-            if (hyperparamIndices[i] + 1 < _hyperParams.get(_hyperParamNames[i]).length) {
-              return true;
-            }
-          }
-          return false;
-        }
-
-        @Override public void reset() {
-          _currentHyperparamIndices = null;
-        }
-
-        @Override
-        public double time_remaining_secs() { return Double.MAX_VALUE; }
-
-        @Override
-        public double max_runtime_secs() { return Double.MAX_VALUE; }
-
-        public int max_models() { return _maxHyperSpaceSize > Integer.MAX_VALUE ? Integer.MAX_VALUE : (int)_maxHyperSpaceSize; }
-
-        @Override
-        public void modelFailed(Model failedModel) {
-          // nada
-        }
-
-        @Override
-        public Object[] getCurrentRawParameters() {
-          Object[] hyperValues = new Object[_hyperParamNames.length];
-          return hypers(_currentHyperparamIndices, hyperValues);
-        }
-      }; // anonymous HyperSpaceIterator class
-    } // iterator()
-
-    /**
-     * Cartesian iteration over the hyper-parameter space, varying one hyperparameter at a
-     * time. Mutates the indices that are passed in and returns them.  Returns NULL when
-     * the entire space has been traversed.
-     */
-    private int[] nextModelIndices(int[] hyperparamIndices) {
-      // Find the next parm to flip
-      int i;
-      for (i = 0; i < hyperparamIndices.length; i++) {
-        if (hyperparamIndices[i] + 1 < _hyperParams.get(_hyperParamNames[i]).length) {
-          break;
-        }
-      }
-      if (i == hyperparamIndices.length) {
-        return null; // All done, report null
-      }
-      // Flip indices
-      for (int j = 0; j < i; j++) {
-        hyperparamIndices[j] = 0;
-      }
-      hyperparamIndices[i]++;
-      return hyperparamIndices;
-    }
-  } // class CartesianWalker
-
-  /**
-   * Hyperparameter space walker which visits random combinations of hyperparameters whose possible values are
-   * given in explicit lists as they are with CartesianWalker.
-   */
-  public static class RandomDiscreteValueWalker<MP extends Model.Parameters>
-      extends BaseWalker<MP, HyperSpaceSearchCriteria.RandomDiscreteValueSearchCriteria> {
-    Random random;
-
-    /** All visited hyper params permutations, including the current one. */
-    private List<int[]> _visitedPermutations = new ArrayList<>();
-    private Set<Integer> _visitedPermutationHashes = new LinkedHashSet<>(); // for fast dupe lookup
-
-    public RandomDiscreteValueWalker(MP params,
-                                     Map<String, Object[]> hyperParams,
-                                     ModelParametersBuilderFactory<MP> paramsBuilderFactory,
-                                     HyperSpaceSearchCriteria.RandomDiscreteValueSearchCriteria search_criteria) {
-      super(params, hyperParams, paramsBuilderFactory, search_criteria);
-
-      if (-1 == search_criteria.seed())
-        random = new Random();                       // true random
-      else
-        random = new Random(search_criteria.seed()); // seeded repeatable pseudorandom
-    }
-
-    /** Based on the last model, the given array of ScoringInfo, and our stopping criteria should we stop early? */
-    @Override
-    public boolean stopEarly(Model model, ScoringInfo[] sk) {
-      return ScoreKeeper.stopEarly(ScoringInfo.scoreKeepers(sk),
-                                   search_criteria().stopping_rounds(),
-                                    ScoreKeeper.ProblemType.forSupervised(model._output.isClassifier()),
-                                   search_criteria().stopping_metric(),
-                                   search_criteria().stopping_tolerance(), "grid's best", true);
-    }
-
-    @Override
-    public HyperSpaceIterator<MP> iterator() {
-      return new HyperSpaceIterator<MP>() {
-        /** Current hyper params permutation. */
-        private int[] _currentHyperparamIndices = null;
-
-        /** One-based count of the permutations we've visited, primarily used as an index into _visitedHyperparamIndices. */
-        private int _currentPermutationNum = 0;
-
-        /** Start time of this grid */
-        private long _start_time = System.currentTimeMillis();
-
-        // TODO: override into a common subclass:
-        @Override
-        public MP nextModelParameters(Model previousModel) {
-          // NOTE: nextModel checks _visitedHyperparamIndices and does not return a duplicate set of indices.
-          // NOTE: in RandomDiscreteValueWalker nextModelIndices() returns a new array each time, rather than
-          // mutating the last one.
-          _currentHyperparamIndices = nextModelIndices();
-
-          if (_currentHyperparamIndices != null) {
-            _visitedPermutations.add(_currentHyperparamIndices);
-            _visitedPermutationHashes.add(integerHash(_currentHyperparamIndices));
-            _currentPermutationNum++; // NOTE: 1-based counting
-
-            // Fill array of hyper-values
-            Object[] hypers = hypers(_currentHyperparamIndices, new Object[_hyperParamNames.length]);
-            // Get clone of parameters
-            MP commonModelParams = (MP) _params.clone();
-            // Fill model parameters
-            MP params = getModelParams(commonModelParams, hypers);
-
-            // add max_runtime_secs in search criteria into params if applicable
-            if (_search_criteria != null && _search_criteria.strategy() == HyperSpaceSearchCriteria.Strategy.RandomDiscrete) {
-              // ToDo: model seed setting will be different for parallel model building.
-              // ToDo: This implementation only works for sequential model building.
-              if (_set_model_seed_from_search_seed) {
-                // set model seed = search_criteria.seed+(0, 1, 2,..., model number)
-                params._seed = _search_criteria.seed() + (model_number++);
-              }
-
-              // set max_runtime_secs
-              double timeleft = this.time_remaining_secs();
-              if (timeleft > 0)  {
-                if (params._max_runtime_secs > 0) {
-                  params._max_runtime_secs = min(params._max_runtime_secs, timeleft);
-                } else {
-                  params._max_runtime_secs = timeleft;
-                }
-              }
-            }
-            return params;
-          } else {
-            throw new NoSuchElementException("No more elements to explore in hyper-space!");
-          }
-        }
-
-        @Override
-        public boolean hasNext(Model previousModel) {
-          // Note: we compare _currentPermutationNum to max_models, because it counts successfully created models, but
-          // we compare _visitedPermutationHashes.size() to _maxHyperSpaceSize because we want to stop when we have attempted each combo.
-          //
-          // _currentPermutationNum is 1-based
-          return (_visitedPermutationHashes.size() < _maxHyperSpaceSize &&
-                  (search_criteria().max_models() == 0 || _currentPermutationNum < search_criteria().max_models())
-          );
-        }
-
-        @Override
-        public void reset() {
-          _start_time = System.currentTimeMillis();
-          _currentPermutationNum = 0;
-          _currentHyperparamIndices = null;
-          _visitedPermutations.clear();
-          _visitedPermutationHashes.clear();
-        }
-
-        public double max_runtime_secs() {
-          return search_criteria().max_runtime_secs();
-        }
-
-        public int max_models() {
-          return search_criteria().max_models();
-        }
-
-        @Override
-        public double time_remaining_secs() {
-          return search_criteria().max_runtime_secs() - (System.currentTimeMillis() - _start_time) / 1000.0;
-        }
-
-        @Override
-        public void modelFailed(Model failedModel) {
-          // Leave _visitedPermutations, _visitedPermutationHashes and _currentHyperparamIndices alone
-          // so we don't revisit bad parameters. Note that if a model build fails for other reasons we
-          // won't retry.
-          _currentPermutationNum--;
-        }
-
-        @Override
-        public Object[] getCurrentRawParameters() {
-          Object[] hyperValues = new Object[_hyperParamNames.length];
-          return hypers(_currentHyperparamIndices, hyperValues);
-        }
-      }; // anonymous HyperSpaceIterator class
-    } // iterator()
-
-    /**
-     * Random iteration over the hyper-parameter space.  Does not repeat
-     * previously-visited combinations.  Returns NULL when we've hit the stopping
-     * criteria.
-     */
-    private int[] nextModelIndices() {
-      int[] hyperparamIndices =  new int[_hyperParamNames.length];
-
-      do {
-        // generate random indices
-        for (int i = 0; i < _hyperParamNames.length; i++) {
-          hyperparamIndices[i] = random.nextInt(_hyperParams.get(_hyperParamNames[i]).length);
-        }
-        // check for aliases and loop if we've visited this combo before
-      } while (_visitedPermutationHashes.contains(integerHash(hyperparamIndices)));
-
-      return hyperparamIndices;
-    } // nextModel
-
-  } // RandomDiscreteValueWalker
 }
